# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * AI-enabled personal finance API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface UserProfile {
  id: string;
  name?: string | null;
  email?: string | null;
  hasConsented: boolean;
  onboardingCompleted: boolean;
  createdAt: string;
}

export interface UpdateProfileRequest {
  name?: string;
}

export interface Account {
  id: number;
  userId: string;
  institutionName: string;
  accountName: string;
  accountType: string;
  balance: string;
  currency: string;
  isActive: boolean;
  createdAt: string;
}

export type TransactionType = typeof TransactionType[keyof typeof TransactionType];


export const TransactionType = {
  debit: 'debit',
  credit: 'credit',
} as const;

export interface Transaction {
  id: number;
  accountId: number;
  categoryId?: number | null;
  amount: string;
  description: string;
  merchantName?: string | null;
  date: string;
  type: TransactionType;
  isRecurring: boolean;
  createdAt: string;
  categoryName?: string | null;
  categoryColor?: string | null;
  categoryIcon?: string | null;
}

export interface TransactionList {
  items: Transaction[];
  total: number;
  limit: number;
  offset: number;
}

export interface Category {
  id: number;
  name: string;
  icon: string;
  color: string;
}

export type GoalStatus = typeof GoalStatus[keyof typeof GoalStatus];


export const GoalStatus = {
  active: 'active',
  completed: 'completed',
  paused: 'paused',
} as const;

export interface Goal {
  id: number;
  userId: string;
  name: string;
  targetAmount: string;
  currentAmount: string;
  targetDate?: string | null;
  category: string;
  status: GoalStatus;
  createdAt: string;
  progressPercent: number;
}

export interface CreateGoalRequest {
  name: string;
  targetAmount: string;
  currentAmount?: string;
  targetDate?: string;
  category?: string;
}

export type UpdateGoalRequestStatus = typeof UpdateGoalRequestStatus[keyof typeof UpdateGoalRequestStatus];


export const UpdateGoalRequestStatus = {
  active: 'active',
  completed: 'completed',
  paused: 'paused',
} as const;

export interface UpdateGoalRequest {
  name?: string;
  targetAmount?: string;
  currentAmount?: string;
  targetDate?: string | null;
  category?: string;
  status?: UpdateGoalRequestStatus;
}

export interface ConsentStatus {
  hasConsented: boolean;
  consentGivenAt?: string | null;
}

export interface OnboardingStatus {
  completed: boolean;
  currentStep: number;
  totalSteps: number;
}

export interface CategorySummary {
  categoryId: number;
  categoryName: string;
  categoryColor: string;
  categoryIcon: string;
  totalAmount: number;
  percentage: number;
}

export interface FinancialSummary {
  totalBalance: number;
  totalIncome: number;
  totalExpenses: number;
  savingsRate: number;
  topCategories: CategorySummary[];
  recentTransactions: Transaction[];
  accountCount: number;
  transactionCount: number;
}

export interface AiTestResponse {
  success: boolean;
  message: string;
  model?: string | null;
}

export interface ApiError {
  error: string;
  message?: string;
}

/**
 * Unauthorized
 */
export type UnauthorizedResponse = ApiError;

/**
 * Not found
 */
export type NotFoundResponse = ApiError;

export type GetTransactionsParams = {
accountId?: number;
categoryId?: number;
startDate?: string;
endDate?: string;
limit?: number;
offset?: number;
};