# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: AI-enabled personal finance API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /goals:
    get:
      operationId: getGoals
      tags: [goals]
      summary: List financial goals
      security:
        - bearerAuth: []
      responses:
        "200":
          description: List of goals
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Goal"
        "401":
          $ref: "#/components/responses/Unauthorized"
    post:
      operationId: createGoal
      tags: [goals]
      summary: Create a financial goal
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateGoalRequest"
      responses:
        "201":
          description: Created goal
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Goal"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /goals/{id}:
    get:
      operationId: getGoal
      tags: [goals]
      summary: Get goal by ID
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Goal detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Goal"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
    put:
      operationId: updateGoal
      tags: [goals]
      summary: Update a financial goal
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateGoalRequest"
      responses:
        "200":
          description: Updated goal
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Goal"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
    delete:
      operationId: deleteGoal
      tags: [goals]
      summary: Delete a financial goal
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted successfully
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
  responses:
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ApiError"
    NotFound:
      description: Not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ApiError"
  schemas:
...
    Goal:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: string
        name:
          type: string
        targetAmount:
          type: string
        currentAmount:
          type: string
        targetDate:
          type: string
          nullable: true
        category:
          type: string
        status:
          type: string
          enum: [active, completed, paused]
        createdAt:
          type: string
          format: date-time
        progressPercent:
          type: number
      required:
        - id
        - userId
        - name
        - targetAmount
        - currentAmount
        - category
        - status
        - createdAt
        - progressPercent
    CreateGoalRequest:
      type: object
      properties:
        name:
          type: string
        targetAmount:
          type: string
        currentAmount:
          type: string
        targetDate:
          type: string
        category:
          type: string
      required:
        - name
        - targetAmount
    UpdateGoalRequest:
      type: object
      properties:
        name:
          type: string
        targetAmount:
          type: string
        currentAmount:
          type: string
        targetDate:
          type: string
          nullable: true
        category:
          type: string
        status:
          type: string
          enum: [active, completed, paused]
...
    ApiError:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
      required:
        - error